+++ southbound/southbound-impl/src/test/java/org/opendaylight/ovsdb/southbound/OvsdbDataTreeChangeListenerTest.java
@@ -30,6 +30,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.ovsdb.re
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.ovsdb.rev150105.ovsdb.node.attributes.ConnectionInfoBuilder;
 import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.network.topology.topology.Node;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint16;
 
 /**
  * Unit tests for the data-tree change listener.
@@ -63,7 +64,7 @@ public class OvsdbDataTreeChangeListenerTest extends AbstractConcurrentDataBroke
         InetAddress inetAddress = InetAddress.getByName("127.0.0.1");
         int port = 6640;
         IpAddress ipAddress = SouthboundMapper.createIpAddress(inetAddress);
-        PortNumber portNumber = new PortNumber(port);
+        PortNumber portNumber = new PortNumber(Uint16.valueOf(port));
 
         final ConnectionInfo connectionInfo = new ConnectionInfoBuilder()
                 .setRemoteIp(ipAddress)
